Split EPL MCP tools into a mounted sub-server (#356)
Move ``get_team_context`` and its seven ESPN fixture helpers
(``_is_final``, ``_is_in_progress``, ``_parse_score``,
``_outcome_for_score``, ``_fixture_to_dict``, ``_result_to_dict``,
``_derive_standings``) from the monolithic ``server.py`` into a new
``odds_mcp.tools.epl`` submodule with its own ``FastMCP("odds-mcp-epl")``
instance. The parent ``mcp`` mounts the child without a namespace, so
``get_team_context`` is exposed verbatim — agent traces continue to call
the tool under the bare name.

Generic tools and helpers (``find_retail_edges``, ``get_match_brief``,
``get_current_odds``, ``_event_to_dict``, ``_DEFAULT_SHARP_BOOKMAKERS``,
etc.) stay in ``server.py``. ``server.py`` shrinks from 1549 to 1257
lines; ``tools/epl.py`` adds 313. Test imports for the moved symbols
are repointed at ``odds_mcp.tools.epl``.

Closes #356.
