+++ packages/odds-mcp/odds_mcp/tools/__init__.py
@@ -1,0 +1,1 @@
+"""Sport-specific MCP sub-servers mounted onto the parent ``odds-mcp`` server."""
